refactor : 기존 코드 삭제
