+++ src/main/java/step2_bak/AutoGame.java
@@ -1,4 +1,4 @@
-package step2;
+package step2_bak;
 
 import java.util.*;
 import java.util.stream.IntStream;
