+++ src/main/java/step2_bak/View/ResultView.java
@@ -1,7 +1,7 @@
-package step2.View;
+package step2_bak.View;
 
-import step2.Data.AnalysisData;
-import step2.Game;
+import step2_bak.Data.AnalysisData;
+import step2_bak.Game;
 
 import java.util.List;
