+++ src/test/java/step2_bak/LottoTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2_bak;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
