+++ src/test/java/step2_bak/TicketTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2_bak;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
